Refactor into more sensible structure and add docstrings
